requirements.txt
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
pytesseract==0.3.10
pdf2image==1.16.3
pypdf2==3.0.1
pillow==10.0.1
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
huggingface_hub==0.19.0